+++ src/cq-tags/CQTag.js
@@ -1,14 +1,30 @@
-/**
- * @abstract
- */
+const CQ = require('./cq-utils')
+
 class CQTag {
-  constructor (type, options = {}) {
+  constructor (type, meta = {}) {
     this.type = type
-    this.options = options
+    this.meta = meta
+  }
+
+  equals (another) {
+    if (typeof another === 'string') {
+      [ another ] = CQ.parse(another)
+    }
+
+    if (!(another instanceof CQTag)) return false
+
+    return this.toString() === another.toString()
+  }
+
+  valueOf () {
+    return this.toString()
   }
 
   toString () {
-    const optStr = Object.keys(this.options).map(k => `${k}=${this.options[k]}`).join(',')
+    const optStr = Object.keys(this.meta)
+      .sort()
+      .map(k => `${k}=${this.meta[k]}`).join(',')
+
     return `[CQ:${this.type}${optStr ? ',' : ''}${optStr}]`
   }
 }